+++ Apizr/Src/Apizr/Logging/HttpMessagePartsExtensions.cs
@@ -1,4 +1,7 @@
-﻿namespace Apizr.Logging;
+﻿using System;
+using Apizr.Configuring.Shared;
+
+namespace Apizr.Logging;
 
 public static class HttpMessagePartsExtensions
 {
@@ -10,4 +13,18 @@ public static class HttpMessagePartsExtensions
     public static bool HasResponseFlags(this HttpMessageParts parts) =>
         parts.HasFlag(HttpMessageParts.ResponseHeaders) ||
         parts.HasFlag(HttpMessageParts.ResponseBody);
+
+    public static Action<T> IgnoreMessageParts<T>(this
+        Action<T> optionsBuilder, HttpMessageParts messageParts)
+        where T : IApizrGlobalSharedOptionsBuilderBase
+    {
+        if (optionsBuilder == null)
+            optionsBuilder = optionsBuilderInstance =>
+                ((IApizrInternalOptionsBuilder)optionsBuilderInstance).SetHandlerParameter(Constants.ApizrIgnoreMessagePartsKey, messageParts);
+        else
+            optionsBuilder += optionsBuilderInstance =>
+                ((IApizrInternalOptionsBuilder)optionsBuilderInstance).SetHandlerParameter(Constants.ApizrIgnoreMessagePartsKey, messageParts);
+
+        return optionsBuilder;
+    }
 }